requirements.txt
django
invoke
black
requests
python-dateutil